{"schema":"decodex/commit/1","summary":"Split Decodex app login sheet components","authority":"XY-1115"}
